Bug fixes to code
Shutdown should not use XOR to disable the shutdown flag.  Doing two
shutdowns in a row would actually start up the chip.
Example code should show correct operation of the chip.  When the chip
is enabled (wakeup) the temperature will take 250ms (in high resolution
mode) to update the temp to the correct value.  The example code only
worked because the device holds the previous time in it's register on
wake, and the delay was AFTER the read.  This caused the temp to be the
LAST temp, not the current temp.